+++ tests/all.ts
@@ -48,6 +48,7 @@ import { testit as issue5840 } from "./issue-5840.ts";
 import { testit as issue5842 } from "./issue-5842.ts";
 import { testit as issue5845 } from "./issue-5845.ts";
 import { testit as issue5846 } from "./issue-5846.ts";
+import { testit as issue5850 } from "./issue-5850.ts";
 import { testit as issue5865 } from "./issue-5865.ts";
 import { testit as parseTipBrackets } from "./parse-tip-brackets.ts";
 import { testit as ghsaP2ph2rvmQ37m } from "./security-ghsa-p2ph-2rvm-q37m.ts";
@@ -92,6 +93,7 @@ const tests: [string, () => void | Promise<void>][] = [
   ["issue-5842", issue5842],
   ["issue-5845", issue5845],
   ["issue-5846", issue5846],
+  ["issue-5850", issue5850],
   ["issue-5865", issue5865],
   ["parse-tip-brackets", parseTipBrackets],
   ["security-ghsa-p2ph-2rvm-q37m", ghsaP2ph2rvmQ37m],